+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Changed
+- Migrate the seminar attachments to FAL (#1067)
 - Migrate the seminar images to FAL (#1065)
 - Migrate the category icons to FAL (#1063)
 - Switch the autoloading to PSR-4 (#1052)